al_rased/requirements.txt
python-telegram-bot[job-queue]
redis
aiosqlite
python-dotenv
pandas
scikit-learn
joblib
